% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BaalChIP-methods.R
\docType{methods}
\name{adjustmentBaalPlot}
\alias{adjustmentBaalPlot}
\title{Plots Allelic Ratios before and after adjustment}
\usage{
\S4method{adjustmentBaalPlot}{BaalChIP}(.Object, col = c("green3", "gray50"))
}
\arguments{
\item{.Object}{An object of the \code{\link{BaalChIP}} class.}

\item{col}{A character vector indicating the colors for the density plot ( default is c( "green3","gray50") )}
}
\description{
Produces a density plot of the distribution of allelic ratios (REF/TOTAL) before and after BaalChIP adjustment for RM and RAF biases.
}
\examples{
data(BaalObject)
adjustmentBaalPlot(BaalObject)
adjustmentBaalPlot(BaalObject, col=alpha(c( "blue","pink")))
}
\author{
Ines de Santiago, Wei Liu, Ke Yuan, Florian Markowetz
}
\seealso{
\code{\link{BaalChIP.report}}, \code{\link{summaryQC}}
#'
}

