% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BaalChIP-methods.R
\docType{methods}
\name{BaalChIP.get}
\alias{BaalChIP.get}
\title{Get slots from a BaalChIP object}
\usage{
\S4method{BaalChIP.get}{BaalChIP}(.Object, what = c("samples", "param",
  "alleleCountsPerBam", "mergedCounts"))
}
\arguments{
\item{.Object}{An object of the \code{\link{BaalChIP}} class}

\item{what}{a single character value specifying which information should be retrieved. Options: 'samples', 'param', 'alleleCountsPerBam', 'mergedCounts'}
}
\value{
The slot content from an object of the \code{\link{BaalChIP}} class.
}
\description{
Get information from individual slots in a BaalChIP object.
}
\examples{
data(BaalObject)

#samples data spreadsheet and hets:
BaalChIP.get(BaalObject,"samples")

#parameters used within run:
BaalChIP.get(BaalObject,"param")

#retrieve a GRanges list with allele-specific read counts per BAM file:
counts <- BaalChIP.get(BaalObject,"alleleCountsPerBam")
counts[["MCF7"]][[1]]

#retrieve a data.frame with allele-specific read counts per group:
counts <- BaalChIP.get(BaalObject,"mergedCounts")
head(counts[[1]])
}
\author{
Ines de Santiago
}

