% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BaalChIP-methods.R
\docType{methods}
\name{BaalChIP.QC}
\alias{BaalChIP.QC}
\title{BaalChIP pipeline - allele counts and QC}
\usage{
\S4method{BaalChIP.QC}{BaalChIP}(.Object)
}
\arguments{
\item{.Object}{An object of the \code{\link{BaalChIP}} class}
}
\value{
An object of the \code{\link{BaalChIP}} class.
}
\description{
BaalChIP.QC is a wrapper convenience function, to compute allele counts and perform quality controls in one step. This function will use the package's defaults.
}
\details{
This function is a wrapper of the following functions: \code{\link{alleleCounts}}, \code{\link{QCfilter}}, \code{\link{mergePerGroup}}, \code{\link{filter1allele}}
}
\examples{
setwd(system.file("test",package="BaalChIP"))
samplesheet <- "example.tsv"
hets <- c("MCF10"="MCF10_hetSNP.txt", "GM12891"="GM12891_hetSNP.txt")
res <- new("BaalChIP", samplesheet=samplesheet, hets=hets)
res <- BaalChIP.QC(res)

#summary of the QC step
summaryQC(res)
}
\author{
Ines de Santiago, Wei Liu, Ke Yuan, Florian Markowetz
}
\seealso{
\code{\link{summaryQC}}, \code{\link{plotQC}}
}

