% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BaalChIP-methods.R
\docType{methods}
\name{BaalChIP.report}
\alias{BaalChIP.report}
\title{Report ASB variants}
\usage{
\S4method{BaalChIP.report}{BaalChIP}(.Object)
}
\arguments{
\item{.Object}{An object of the \code{\link{BaalChIP}} class}
}
\value{
A named list, with a data.frame per group.
}
\description{
Generates a data.frame per group with significant allele-specific binding (ASB) variants
}
\details{
The reported data frame contains the following columns:
\itemize{
\item ID: unique identifier string per analysed variant
\item Bayes_lower: Lower interval for the estimated allelic ratio (allelic ratio is given by REF / TOTAL)
\item Bayes_upper: Uper interval for the estimated allelic ratio (allelic ratio is given by REF / TOTAL)
\item Bayes_sig_A: classification of variants into allele-specific towards the reference allele. 1/0 means TRUE/FALSE
\item Bayes_sig_B: classification of variants into allele-specific towards the non-reference allele. 1/0 means TRUE/FALSE
}
}
\examples{
data(BaalObject)
report <- BaalChIP.report(BaalObject)

#the reported list is grouped by group_name:
names(report)

#check out the report for one of the groups:
head(report[["MCF10"]])
}
\author{
Ines de Santiago, Wei Liu, Ke Yuan, Florian Markowetz
}
\seealso{
\code{\link{summaryASB}}, \code{\link{getASB}}
}

