% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaalChIP-generics.R
\docType{data}
\name{blacklist_hg19}
\alias{blacklist_hg19}
\title{Blacklisted genomic regions}
\description{
Blacklisted regions downloaded from the UCSC Genome Browser (mappability track; hg19, wgEncodeDacMapabilityConsensusExcludable and wgEncodeDukeMapabilityRegionsExcludable tables). Used as 'RegionsToFilter' within the QCfilter function so that variants overlapping these regions will be removed.
}
\author{
Ines de Santiago \email{ines.desantiago@cruk.cam.ac.uk}
}
\keyword{blacklist_hg19}

