% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaalChIP-generics.R
\docType{data}
\name{ENCODEexample}
\alias{ENCODEexample}
\title{ENCODEexample example dataset. This is a BaalChIP-class object}
\description{
ENCODEexample example dataset. This is a BaalChIP-class object
}
\author{
Ines de Santiago \email{ines.desantiago@cruk.cam.ac.uk}
}
\keyword{ENCODEexample}

