% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BaalChIP-methods.R
\docType{methods}
\name{filterIntbias}
\alias{filterIntbias}
\title{Removes variants for which simulated single-end reads don't align correctly to the original simulated position.}
\usage{
\S4method{filterIntbias}{BaalChIP}(.Object, simul_output = NULL,
  tmpfile_prefix = NULL, simulation_script = "local")
}
\arguments{
\item{.Object}{An object of the \code{\link{BaalChIP}} class}

\item{simul_output}{a non-empty character vector giving the directory of where to save the FASTQ and BAM files generated by the function. If NULL, a random directory under the current working directory will be generated.}

\item{tmpfile_prefix}{an optional character vector giving the initial part of the name of the FASTQ and BAM files generated by the function. If NULL, a random name will be generated.}

\item{simulation_script}{the file path for simulation script containing the instructions of simulation and alignment commands. If NULL, the default simulation script distributed with BaalChIP ('extra/simulation_run.sh') will be used.}
}
\value{
An updated \code{\link{BaalChIP}} object with the slot \code{alleleCounts} containing a list of GRanges objects that pass filters.
}
\description{
Filters the data frame available within a \code{\link{BaalChIP}} object (slot \code{alleleCounts}). This filter performs simulations of reads of the same length as the original ChIP-seq reads, aligns the simulated reads to the genome, calculates the allelic ratios for each variant and finally and ignores those variants for which the allelic ratio (REF/TOTAL) is different than 0.5.
}
\examples{
setwd(system.file("test",package="BaalChIP"))
samplesheet <- "exampleChIP.tsv"
hets <- c("MCF7"="MCF7_hetSNP.txt", "GM12891"="GM12891_hetSNP.txt")
res <- new("BaalChIP", samplesheet=samplesheet, hets=hets)
res <- alleleCounts(res, min_base_quality=10, min_mapq=15)
testingScript=TRUE #For demonstration purposes only (read details in vignette)
res <- filterIntbias(res, simul_output=system.file("test/simuloutput",package="BaalChIP"), tmpfile_prefix="c67c6ec6c433")

#check results
plotSimul(res)
summaryQC(res)
}
\author{
Ines de Santiago
}
\seealso{
\code{\link{BaalChIP.get}}, \code{\link{plotQC}}, \code{\link{summaryQC}}
}

