% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BaalChIP-methods.R
\docType{methods}
\name{getASB}
\alias{getASB}
\title{Identifies allele-specific binding events}
\usage{
\S4method{getASB}{BaalChIP}(.Object, Iter = 5000, conf_level = 0.95,
  RMcorrection = TRUE, RAFcorrection = TRUE)
}
\arguments{
\item{.Object}{An object of the \code{\link{BaalChIP}} class.}

\item{Iter}{Maximun number of iterations (default 5000)}

\item{conf_level}{Confidence interval in the estimated allelic ratio (default 0.95)}

\item{RMcorrection}{Logical value indicating if reference mapping (RM) bias should be applied (default TRUE). If FALSE will not correct for reference allele mapping bias. If TRUE will estimate the RM bias from the overall reference allele proportion}

\item{RAFcorrecion}{Logical value indicating if relative allele frequency (RAF) bias correction should be applyed (default TRUE). If TRUE will read RAF values for each variant from \code{hets} files (RAF column name). If FALSE will not correct for relative allele frequency bias}
}
\value{
An updated \code{\link{BaalChIP}} object with the slot \code{ASB} containing variants identified as allele-specific.
}
\description{
getASB identifies allele-specific binding events using a bayesian framework.
}
\examples{
setwd(system.file("test",package="BaalChIP"))
samplesheet <- "example.tsv"
hets <- c("MCF10"="MCF10_hetSNP.txt", "GM12891"="GM12891_hetSNP.txt")
res <- new("BaalChIP", samplesheet=samplesheet, hets=hets)
res <- BaalChIP.QC(res)
res <- getASB(res)

#summary - number of significant ASB variants
summaryASB(res)

#report result
res <- BaalChIP.report(res)
}
\author{
Ines de Santiago, Wei Liu, Ke Yuan, Florian Markowetz
}
\seealso{
\code{\link{summaryASB}}, \code{\link{BaalChIP.report}}
}

