% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaalChIP-generics.R
\docType{data}
\name{pickrell2011cov1_hg19}
\alias{pickrell2011cov1_hg19}
\title{Genomic regions of collapsed repeats}
\description{
Collapsed repeat regions at the 0.1\% threshold (hg19 reference). Used as 'RegionsToFilter' within the QCfilter function so that variants overlapping these regions will be removed.
}
\author{
Ines de Santiago \email{ines.desantiago@cruk.cam.ac.uk}
}
\references{
Pickrell et al., 2011 (\url{http://www.ncbi.nlm.nih.gov/pubmed/21690102})
}
\keyword{pickrell2011cov1_hg19}

