% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BaalChIP-methods.R
\docType{methods}
\name{plotQC}
\alias{plotQC}
\title{Plots QC results}
\usage{
\S4method{plotQC}{BaalChIP}(.Object, what = "barplot_per_group",
  addlegend = TRUE, plot = TRUE)
}
\arguments{
\item{.Object}{An object of the \code{\link{BaalChIP}} class.}

\item{what}{A single character value indicating the type of plot. Options:
\itemize{
\item \code{overall_pie}: plots the average percentage of variants in each filter category (averaged across all groups analysed)
\item \code{boxplot_per_filter}: plots the number of variants that were filtered out per filter category
\item \code{barplot_per_group}: plots the number of variants that were filtered out per group
}}

\item{addlegend}{A logical value indicating if legend should be included in the plot (default TRUE)}
}
\description{
Produces different plots of QC results.
}
\examples{
data(BaalObject)
plotQC(BaalObject, "overall_pie")
plotQC(BaalObject, "boxplot_per_filter", addlegend=FALSE)
plotQC(BaalObject, "barplot_per_group")
}
\author{
Ines de Santiago, Wei Liu, Ke Yuan, Florian Markowetz
}
\seealso{
\code{\link{BaalChIP.run}}, \code{\link{summaryQC}}
#'
}

