% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaalChIP-methods.R
\docType{methods}
\name{plotSimul}
\alias{plotSimul}
\title{Plots the percentage of correctly aligned simulated reads}
\usage{
\S4method{plotSimul}{BaalChIP}(.Object, plot = TRUE)
}
\arguments{
\item{.Object}{An object of the \code{\link{BaalChIP}} class.
#' @examples
data(BaalObject)
plotSimul(BaalObject)}
}
\description{
Produces a plot of the proportion of variants that displayed the correct number of mapped simulated reads.
}
\author{
Ines de Santiago
}

