% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaalChIP-methods.R
\docType{methods}
\name{QCfilter}
\alias{QCfilter}
\title{Removes variants that may be problematic for identification of allele-specific events}
\usage{
\S4method{QCfilter}{BaalChIP}(.Object, RegionsToFilter = NULL,
  RegionsToKeep = NULL)
}
\arguments{
\item{.Object}{An object of the \code{\link{BaalChIP}} class}

\item{RegionsToFilter}{a named list of GRanges objects. Variants overlapping these regions will be removed}

\item{RegionsToKeep}{a named list of GRanges objects. Works in an oposite way to 'RegionstoFilter', variants NOT overlapping these regions will be removed}
}
\value{
An updataded \code{\link{BaalChIP}} object with the slot \code{alleleCounts} containing a list of GRanges objects that pass filters.
}
\description{
Quality control step for removing variants that may be problematic for identification of allele-specific events.
}
\examples{
setwd(system.file("test",package="BaalChIP"))
samplesheet <- "exampleChIP.tsv"
hets <- c("MCF7"="MCF7_hetSNP.txt", "GM12891"="GM12891_hetSNP.txt")
res <- new("BaalChIP", samplesheet=samplesheet, hets=hets)
res <- alleleCounts(res, min_base_quality=10, min_mapq=15)
data(blacklist_hg19)
data(pickrell2011cov1_hg19)
data(UniqueMappability50bp_hg19)
res <- QCfilter(res,
               RegionsToFilter=list("blacklist"=blacklist_hg19, "highcoverage"=pickrell2011cov1_hg19),
               RegionsToKeep=list("UniqueMappability"=UniqueMappability50bp_hg19))

#check results
plotQC(res,"barplot")
summaryQC(res)
}
\author{
Ines de Santiago
}
\seealso{
\code{\link{BaalChIP.get}}, \code{\link{plotQC}}, \code{\link{summaryQC}}
}

