% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaalChIP-methods.R
\docType{methods}
\name{summaryASB}
\alias{summaryASB}
\title{Summary of ASB test}
\usage{
\S4method{summaryASB}{BaalChIP}(.Object)
}
\arguments{
\item{.Object}{An object of the \code{\link{BaalChIP}} class}
}
\value{
A matrix containning the total number of allele-specific variants (TOTAL) and the number of variants allele-specific for the reference (REF) and alternate alleles (ALT).
}
\description{
Generates summary of ASB test result.
}
\examples{
data(BaalObject)
summaryASB(BaalObject)
}
\author{
Ines de Santiago
}
\seealso{
\code{\link{getASB}}, \code{\link{BaalChIP.report}}
}

