% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BaalChIP-methods.R
\docType{methods}
\name{summaryQC}
\alias{summaryQC}
\title{Summary of QC}
\usage{
\S4method{summaryQC}{BaalChIP}(.Object)
}
\arguments{
\item{.Object}{An object of the \code{\link{BaalChIP}} class}
}
\value{
A list with two elements:
\itemize{
\item \code{filtering_stats} containning the number of variants that were filtered out in each filter category and the total number that 'pass' all filters
\item \code{average_stats} containning the average number and average percentage of variants in each filter category, averaged across all analysed groups
}
}
\description{
Generates summary of QC result.
}
\examples{
data(BaalObject)
summaryQC(BaalObject)
}
\author{
Ines de Santiago, Wei Liu, Ke Yuan, Florian Markowetz
}
\seealso{
\code{\link{BaalChIP.QC}}, \code{\link{plotQC}}
}

