% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaalChIP-generics.R
\docType{data}
\name{UniqueMappability50bp_hg19}
\alias{UniqueMappability50bp_hg19}
\title{Genomic regions of unique mappability}
\description{
Unique regions with genomic mappability score equal to 1 selected from DUKE uniqueness mappability track of the UCSC genome browser generated using a window size of 50bp (hg19,  wgEncodeCrgMapabilityAlign50mer table). Used as 'RegionsToKeep' within the QCfilter function so that variants NOT overlapping these regions will be removed.
}
\author{
Ines de Santiago \email{ines.desantiago@cruk.cam.ac.uk}
}
\keyword{UniqueMappability50bp_hg19}

