% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaalChIP-generics.R
\docType{data}
\name{BaalObject}
\alias{BaalObject}
\title{BaalObject example dataset}
<<<<<<< HEAD
\format{
A BaalChIP-class object
}
=======
\format{A BaalChIP-class object}
>>>>>>> upstream/master
\description{
BaalObject example dataset
}
\author{
Ines de Santiago \email{ines.desantiago@cruk.cam.ac.uk}
}
\keyword{BaalObject}
<<<<<<< HEAD
=======

>>>>>>> upstream/master
