% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaalChIP-generics.R
\docType{data}
\name{ENCODEexample}
\alias{ENCODEexample}
\title{ENCODEexample example dataset}
\format{
A BaalChIP-class object
}
\source{
Downloaded from supplementary data of BaalChIP paper
}
\description{
ENCODEexample example dataset
}
\author{
Ines de Santiago \email{ines.desantiago@cruk.cam.ac.uk}
}
\keyword{ENCODEexample}
