% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaalChIP-generics.R
\docType{data}
\name{FAIREexample}
\alias{FAIREexample}
\title{FAIREexample example dataset}
<<<<<<< HEAD
\format{
A BaalChIP-class object
}
=======
\format{A BaalChIP-class object}
>>>>>>> upstream/master
\source{
Downloaded from supplementary data of BaalChIP paper
}
\description{
FAIREexample example dataset
}
\author{
Ines de Santiago \email{ines.desantiago@cruk.cam.ac.uk}
}
\keyword{FAIREexample}
<<<<<<< HEAD
=======

>>>>>>> upstream/master
