% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaalChIP-generics.R, R/BaalChIP-methods.R
\docType{methods}
\name{plotSimul}
\alias{plotSimul}
\alias{plotSimul,BaalChIP-method}
\title{Method plotSimul}
\usage{
plotSimul(.Object, plot = TRUE)

\S4method{plotSimul}{BaalChIP}(.Object, plot = TRUE)
}
\arguments{
\item{.Object}{An object of the \code{\link{BaalChIP}} class.}

\item{plot}{a logical value to whether it should plot (TRUE) or not (FALSE). Default is TRUE.}
}
\value{
A plot
}
\description{
Method plotSimul

Produces a plot of the proportion of variants that displayed the correct number of mapped simulated reads.
}
\examples{
data('BaalObject')
plotSimul(BaalObject)
}
\author{
Ines de Santiago
}

