% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaalChIP-generics.R, R/BaalChIP-methods.R
\name{summaryASB}
\alias{summaryASB}
\alias{summaryASB,BaalChIP-method}
\title{Method summaryASB}
\usage{
summaryASB(.Object)

\S4method{summaryASB}{BaalChIP}(.Object)
}
\arguments{
\item{.Object}{An object of the \code{\link{BaalChIP}} class.}
}
\value{
A matrix containing the total number of allele-specific variants (TOTAL) and the number of variants allele-specific for the reference (REF) and alternate alleles (ALT).
}
\description{
Generates summary of ASB test result.
}
\examples{
data('BaalObject')
summaryASB(BaalObject)
}
\seealso{
\code{\link{getASB}}, \code{\link{BaalChIP.report}}
}
\author{
Ines de Santiago
}
